common/prefill: remove standalone serving mode
Request + producer is the only serving path, so the bounded golden-trace loop and
its per-rank chunk-count agreement (PREFILL_STANDALONE_NCHUNKS in every binding's
global_env) have no caller left. The per-user KV cache is now sized from
PREFILL_MAX_SEQ_LEN alone, defaulting to 11 chunks.
