+++ models/demos/common/prefill/runners/prefill_runner.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
 
-"""Disaggregated prefill runner — one entry point, two run modes that share the same N-rank pipeline.
+"""Disaggregated prefill runner — one entry point driving an N-rank serving pipeline.
 
 Model-agnostic: the model is selected by PREFILL_MODEL and driven through a PrefillModelAdapter
 (see ../adapter.py and ADDING_A_PREFILL_MODEL.md). This driver wires rank topology, input,
@@ -13,19 +13,17 @@ The model is split across N ranks under tt-run: each rank owns a contiguous laye
 the same TtPrefillRuntime (first_layer_idx / is_first_rank / is_last_rank). With >1 rank the cross-rank
 hidden state moves device-to-device over fabric sockets (connected MGD + FABRIC_2D); N=1 is the
 single-galaxy case (no transport). Ranks run decoupled (no per-chunk barrier; one warm-up barrier
-after compile). The two modes run identical pipeline mechanics and differ only in the trigger:
+after compile).
 
-  * Request mode (default): production serving. rank 0's tokens + per-iter PrefillMetadata arrive over
-    the H2D socket from an external producer (prefill_producer.py / the scheduler); the loop is
-    UNBOUNDED. KV-chunk-table migration + per-layer LayerAck are wired for the single-rank case only
-    (disabled for the pipeline for now). Shutdown is graceful: the producer/scheduler closes the stream
-    with an all -1 PrefillMetadata sentinel that each rank forwards downstream and then exits on; a rank
-    blocked in the recv can only be released by a transfer (the recv device op has no timeout), so
-    SIGTERM/SIGKILL remains the hard fallback if no sentinel arrives.
-
-  * Standalone mode (PREFILL_STANDALONE=1): bring-up / benchmark. rank 0's input is the golden trace
-    for a fixed PREFILL_STANDALONE_NCHUNKS chunks; the loop is BOUNDED and exits cleanly.
-    PREFILL_STANDALONE_PCC=1 checks each rank's KV slice vs the golden.
+Serving is request-driven: rank 0's tokens + per-iter PrefillMetadata arrive over the H2D socket from
+an external producer (prefill_producer.py / the scheduler); the loop is UNBOUNDED. KV-chunk-table
+migration and per-layer LayerAck run at any rank count: every rank joins the all-gather that merges
+the chunk table and rank 0 publishes it, and pipeline layer completions are routed to the master
+rank, which re-emits them into the same ack channel the scheduler connects to in the single-rank case
+(only PREFILL_MOCK_MIGRATION stays single-rank). Shutdown is graceful: the producer/scheduler closes
+the stream with an all -1 PrefillMetadata sentinel that each rank forwards downstream and then exits
+on; a rank blocked in the recv can only be released by a transfer (the recv device op has no timeout),
+so SIGTERM/SIGKILL remains the hard fallback if no sentinel arrives.
 
 The model class is the single source of truth — this driver wires rank topology, input, transport,
 and the per-chunk schedule; it does not reimplement embed / layers / forward.
@@ -45,9 +43,8 @@ from models.demos.common.prefill.runners.migration import serialize_device_map
 from models.demos.common.prefill.runners.runner_utils import (
     activation_global_spec,
     build_h2d_service,
-    load_trace_token_ids,
+    compute_layer_split,
     open_mesh_device,
-    resolve_trace_dir,
 )
 
 # NOTE: the layer_completion classes (the standalone `_layer_completion` extension)
@@ -168,11 +165,13 @@ _tp = int(os.environ.get("PREFILL_TP", 4))
 GLOBAL_MESH_SHAPE = (_sp, _tp)
 NUM_LAYERS = int(os.environ.get("PREFILL_NUM_LAYERS", 61))
 CHUNK_SIZE = int(os.environ.get("PREFILL_CHUNK_SIZE", 5 * 1024))
-# Chunks this run drives. The per-user KV cache is sized to exactly hold them
-# (max_seq_len = chunk_size * num_chunks), so there is no separate cache-length knob to keep in sync.
-# PREFILL_MAX_SEQ_LEN still overrides if a larger cache is wanted.
-NUM_CHUNKS = int(os.environ.get("PREFILL_STANDALONE_NCHUNKS", 11))
-MAX_SEQ_LEN = int(os.environ.get("PREFILL_MAX_SEQ_LEN", CHUNK_SIZE * NUM_CHUNKS))
+# Per-user KV cache length. In request mode the external producer decides the chunk count, so this is
+# the one cache-sizing knob; a chunk must not push a slot past it. Default holds 11 chunks.
+MAX_SEQ_LEN = int(os.environ.get("PREFILL_MAX_SEQ_LEN", CHUNK_SIZE * 11))
+# Chunks one slot's cache holds. Only the migration self-test needs a chunk count (it bounds the
+# otherwise-unbounded loop so the post-loop verify can run); the producer fills a slot to its cache
+# depth there, so the two agree by construction.
+CHUNKS_PER_SLOT = MAX_SEQ_LEN // CHUNK_SIZE
 NUM_USERS = int(os.environ.get("PREFILL_NUM_USERS", 2))
 CAPACITY_FACTOR = int(os.environ.get("PREFILL_CAPACITY_FACTOR", 8))
 _gate_mode_name = os.environ.get("PREFILL_GATE_FALLBACK_MODE", ADAPTER.default_gate_mode)
@@ -200,100 +199,6 @@ _shutdown = False
 def _handle_sigterm(signum, frame):
     global _shutdown
     _shutdown = True
-
-
-# ---------------------------------------------------------------------------
-# Layer assignment
-# ---------------------------------------------------------------------------
-
-
-def _snap_counts_to_starts(counts, valid_starts, num_layers):
-    """Nudge an even split's interior rank boundaries onto the nearest valid start (preserving
-    sum == num_layers), for models that constrain where a rank may begin (layer_split_boundaries).
-    Nearest by |distance| then lower index; each boundary is used at most once and stays increasing."""
-    valid = sorted(valid_starts)
-    boundaries, s = [], 0
-    for c in counts[:-1]:
-        s += c
-        boundaries.append(s)
-    snapped, prev = [], 0
-    for b in boundaries:
-        cand = min(
-            (v for v in valid if prev < v < num_layers and v not in snapped),
-            key=lambda v: (abs(v - b), v),
-            default=None,
-        )
-        if cand is None:
-            raise ValueError(f"cannot place {len(counts)} pipeline ranks on valid layer boundaries {valid}")
-        snapped.append(cand)
-        prev = cand
-    out, prev = [], 0
-    for b in [*snapped, num_layers]:
-        out.append(b - prev)
-        prev = b
-    return out
-
-
-def compute_layer_split(num_layers: int, num_ranks: int, valid_starts=None) -> list[tuple[int, int]]:
-    """Contiguous (first_layer_idx, count) per rank. PREFILL_PP_LAYER_COUNTS, a
-    comma-separated count list summing to num_layers, overrides the default even
-    split (remainder handed to the earlier ranks).
-
-    ``valid_starts`` (from the adapter's ``layer_split_boundaries``): layer indices at which a rank may
-    begin. None => unconstrained. When set, the default even split is auto-snapped onto valid
-    boundaries, and any split (explicit or snapped) whose rank starts fall off them is rejected early."""
-    override = os.environ.get("PREFILL_PP_LAYER_COUNTS")
-    if override:
-        counts = [int(x) for x in override.split(",")]
-        if len(counts) != num_ranks or sum(counts) != num_layers:
-            raise ValueError(
-                f"PREFILL_PP_LAYER_COUNTS={override!r} must list {num_ranks} counts summing to "
-                f"{num_layers} (got {len(counts)} counts summing to {sum(counts)})"
-            )
-    else:
-        base, rem = divmod(num_layers, num_ranks)
-        counts = [base + (1 if r < rem else 0) for r in range(num_ranks)]
-        if valid_starts is not None:
-            counts = _snap_counts_to_starts(counts, valid_starts, num_layers)
-
-    ranges = []
-    start = 0
-    for count in counts:
-        ranges.append((start, count))
-        start += count
-
-    if valid_starts is not None:
-        for first_idx, _ in ranges:
-            if first_idx not in valid_starts:
-                near = sorted(b for b in valid_starts if abs(b - first_idx) <= 4)
-                raise ValueError(
-                    f"pipeline rank starts at layer {first_idx}, not a valid boundary for this model "
-                    f"(nearest valid: {near}). Set PREFILL_PP_LAYER_COUNTS so every cumulative boundary "
-                    f"is a valid start."
-                )
-    return ranges
-
-
-# ---------------------------------------------------------------------------
-# Input
-# ---------------------------------------------------------------------------
-
-
-def _load_token_ids() -> list[int]:
-    """Load this run's token IDs (same source as the single-rank standalone loop).
-    All ranks load identically so they agree on the chunk schedule."""
-    import json
-
-    trace_dir = resolve_trace_dir(os.environ.get("PREFILL_TRACE_DIR", ADAPTER.prefill_trace_default))
-    input_override = os.environ.get("PREFILL_STANDALONE_INPUT")
-    if input_override:
-        with open(input_override) as f:
-            token_ids = list(json.load(f)["token_ids"])
-        logger.info(f"[pp] input override: {len(token_ids)} token_ids from {input_override}")
-    else:
-        logger.info(f"[pp] reading input token_ids from {trace_dir}/metadata.json")
-        token_ids = load_trace_token_ids(trace_dir)
-    return token_ids
 
 
 # ---------------------------------------------------------------------------
@@ -373,13 +278,6 @@ def build_layer_completion_sink(producer, *, source_rank, num_layers):
 # ---------------------------------------------------------------------------
 # Loop
 # ---------------------------------------------------------------------------
-
-
-def _first_rank_chunk_tokens(runtime, token_ids: list[int], kv_actual: int) -> ttnn.Tensor:
-    """Slice this chunk's tokens and build the SP-sharded input tensor. Delegates to the runtime's own
-    builder so the input format has one source of truth."""
-    cfg = runtime.config
-    return runtime.make_chunk_input(token_ids[kv_actual : kv_actual + cfg.chunk_size])
 
 
 def _is_shutdown_sentinel(meta: dict) -> bool:
@@ -611,8 +509,7 @@ def run_request_loop(
     """Production serving loop — UNBOUNDED. rank 0 reads each chunk from the H2D socket (the external
     producer decides the count); downstream ranks read from D2D. Runs until the producer/scheduler
     closes the stream with the all -1 shutdown sentinel (each rank forwards it and exits gracefully) or,
-    as a hard fallback, until SIGTERM/SIGKILL. No fixed NUM_CHUNKS bound, no trace input — see
-    run_standalone_loop for the bounded/trace variant.
+    as a hard fallback, until SIGTERM/SIGKILL. No fixed chunk bound, no trace input, no PCC.
 
     Exception: in migration-validation mode (PREFILL_VALIDATE_MIGRATION=1) the scheduler driver never
     pushes the shutdown sentinel — it pushes PREFILL_STANDALONE_CHUNKED_NCHUNKS chunks, migrates, then
@@ -629,12 +526,10 @@ def run_request_loop(
         f"[pp rank {rank}/{num_ranks}] request (unbounded) loop start "
         f"(is_first={cfg.is_first_rank} is_last={cfg.is_last_rank} input={'h2d' if cfg.is_first_rank else 'd2d'})"
     )
-    # Self-test bound: PREFILL_MIGRATION_SELFTEST=1 makes the loop run exactly NUM_CHUNKS chunks then
-    # exit CLEANLY so the post-loop migrate + verify can run — without it the unbounded loop blocks in
-    # recv and only SIGKILL exits, which kills before the verify. NUM_CHUNKS is the run's single chunk
-    # count (the per-user KV cache is sized to exactly hold it, max_seq_len = chunk_size * NUM_CHUNKS),
-    # and the producer pushes the same count, so they match by construction. 0 == unbounded serving.
-    n_selftest = NUM_CHUNKS if os.environ.get("PREFILL_MIGRATION_SELFTEST", "0") == "1" else 0
+    # Self-test bound: PREFILL_MIGRATION_SELFTEST=1 makes the loop run exactly CHUNKS_PER_SLOT chunks
+    # then exit CLEANLY so the post-loop migrate + verify can run — without it the unbounded loop blocks
+    # in recv and only SIGKILL exits, which kills before the verify. 0 == unbounded serving.
+    n_selftest = CHUNKS_PER_SLOT if os.environ.get("PREFILL_MIGRATION_SELFTEST", "0") == "1" else 0
     t0 = time.perf_counter()
     c = 0
     first = None
@@ -643,6 +538,7 @@ def run_request_loop(
     chunks_per_slot: dict = {}
     real_end_per_slot: dict = {}
     # If we run prefill validation, we need to know the expected number of chunks to exit the loop.
+    # PREFILL_STANDALONE_CHUNKED_* is migration-validation config, not a serving-mode knob.
     _expected_chunks = (
         int(os.environ.get("PREFILL_STANDALONE_CHUNKED_NCHUNKS", "0"))
         if os.environ.get("PREFILL_VALIDATE_MIGRATION", "0") == "1"
@@ -726,78 +622,6 @@ def run_request_loop(
     return chunks_per_slot, real_end_per_slot, c
 
 
-def run_standalone_loop(runtime, kv_caches, rank: int, num_ranks: int, *, d2d_in=None, d2d_out=None) -> None:
-    """Bring-up / benchmark loop — BOUNDED, golden-trace input. rank 0 drives NUM_CHUNKS chunks from the
-    trace; downstream ranks receive the same count over D2D. Every rank knows NUM_CHUNKS (propagated via
-    global_env), so each loops a fixed range independently — no end-of-stream marker needed. With
-    PREFILL_STANDALONE_PCC=1 each rank checks the KV slice it populated vs the golden trace."""
-    cfg = runtime.config
-    slot_id = 0  # first rank fills slot 0; downstream ranks adopt the slot from the received metadata
-    n_chunks = NUM_CHUNKS
-    token_ids = None
-    if cfg.is_first_rank:
-        token_ids = _load_token_ids()
-        token_ids = (token_ids + [1] * (n_chunks * cfg.chunk_size))[: n_chunks * cfg.chunk_size]
-        if n_chunks * cfg.chunk_size > cfg.max_seq_len:
-            raise ValueError(
-                f"{n_chunks} chunks x {cfg.chunk_size} exceeds per-user cache max_seq_len={cfg.max_seq_len}; "
-                f"raise PREFILL_MAX_SEQ_LEN."
-            )
-    # Every rank loops a fixed range(n_chunks) independently — there is no end-of-stream marker, so all
-    # ranks MUST resolve the same PREFILL_STANDALONE_NCHUNKS (set in the binding's global_env, not a
-    # per-rank override). A mismatch strands the pipeline: a low downstream count exits early and leaves
-    # rank 0's next send unconsumed. Log each rank's count so a mismatch is visible across the tag logs.
-    logger.info(
-        f"[pp rank {rank}/{num_ranks}] standalone (bounded) loop start "
-        f"(is_first={cfg.is_first_rank} is_last={cfg.is_last_rank} input=trace chunks={n_chunks})"
-    )
-    t0 = time.perf_counter()
-    first = None
-    for c in range(n_chunks):
-        _lease_reclaim(d2d_in, d2d_out)
-        if cfg.is_first_rank:
-            kv_actual = c * cfg.chunk_size
-            inp = _first_rank_chunk_tokens(runtime, token_ids, kv_actual)
-            meta = {"slot_id": slot_id, "actual_start": kv_actual, "actual_end": kv_actual + cfg.chunk_size}
-        else:
-            # The received metadata tensor is unused here: standalone emits no layer acks. The ack record
-            # is a per-chunk socket metadata tensor, and rank 0 synthesizes its chunks locally (no inbound
-            # socket, so no record) — single-rank standalone IS rank 0, so there is nothing to ack with.
-            # The D2H ack path is wired in _serve_request only; PREFILL_ENABLE_LAYER_ACK is not read here.
-            inp, meta, _ = _d2d_recv(d2d_in)
-            slot_id = meta["slot_id"]
-        t = _compute_and_send(runtime, kv_caches, rank, c, inp, meta, d2d_out)
-        if first is None:
-            first = t
-    # Every rank must finish receiving + forwarding the final chunk before any rank reclaims its
-    # outbound fabric link in the drain. Without this, the producer reclaims the shared link
-    # (share_fabric_links) right after its last send and strands the downstream's final recv —
-    # the pipeline tail deadlocks (ranks 2/3 hang on the last chunk).
-    if num_ranks > 1:
-        ttnn.distributed_context_barrier()
-    _drain_and_log_e2e(runtime, rank, d2d_out, first, n_chunks, t0)
-
-    if os.environ.get("PREFILL_STANDALONE_PCC", "0") == "1":
-        # Each rank PCC-checks the KV slice it populated against the golden trace (offset by
-        # first_layer_idx); all ranks passing == the rank-sliced model reproduces single-rank KV.
-        # kv_cache_pcc_check is an OPTIONAL runtime hook (golden-trace bring-up only — never used in
-        # production serving), so a model whose runtime doesn't implement it can't be checked this way.
-        pcc_check = getattr(runtime, "kv_cache_pcc_check", None)
-        if pcc_check is None:
-            raise RuntimeError(
-                f"PREFILL_STANDALONE_PCC=1 but {type(runtime).__name__} implements no kv_cache_pcc_check "
-                "(optional bring-up hook; see ADDING_A_PREFILL_MODEL.md §2)."
-            )
-        # Pass the raw trace path; the validation helper resolves it (descends the vllm hash subdir).
-        pcc_check(
-            kv_caches,
-            slot_id=slot_id,
-            n_chunks=n_chunks,
-            trace_dir=os.environ.get("PREFILL_TRACE_DIR", ADAPTER.prefill_trace_default),
-            first_layer_idx=cfg.first_layer_idx,
-        )
-
-
 # ---------------------------------------------------------------------------
 # Entry point
 # ---------------------------------------------------------------------------
@@ -819,18 +643,14 @@ def _print_config() -> None:
         ("PREFILL_KV_ONLY_LAST_LAYER", str(KV_ONLY_LAST_LAYER)),
         ("PREFILL_USE_TRACE", f"{USE_TRACE} (trace_region={_TRACE_REGION_SIZE >> 20} MB)"),
         ("PREFILL_CHUNK_SIZE", str(CHUNK_SIZE)),
-        ("PREFILL_STANDALONE_NCHUNKS", str(NUM_CHUNKS)),
         ("PREFILL_MAX_SEQ_LEN", str(MAX_SEQ_LEN)),
         ("PREFILL_NUM_USERS", str(NUM_USERS)),
         ("PREFILL_CAPACITY_FACTOR", str(CAPACITY_FACTOR)),
         ("PREFILL_GATE_FALLBACK_MODE", _gate_mode_name),
         ("PREFILL_FABRIC_MODE", os.environ.get("PREFILL_FABRIC_MODE", "<auto: 1d if sp<=8 else 2d>")),
-        ("PREFILL_STANDALONE (pipeline/bring-up mode)", os.environ.get("PREFILL_STANDALONE", "0")),
         ("PREFILL_PP_D2D_FIFO_BYTES", str(D2D_FIFO_SIZE_BYTES)),
         ("PREFILL_H2D_SERVICE_ID", os.environ.get("PREFILL_H2D_SERVICE_ID", "ds_prefill")),
         ("PREFILL_TRACE_DIR", os.environ.get("PREFILL_TRACE_DIR", ADAPTER.prefill_trace_default)),
-        ("PREFILL_STANDALONE_INPUT", os.environ.get("PREFILL_STANDALONE_INPUT", "<trace default>")),
-        ("PREFILL_STANDALONE_PCC", os.environ.get("PREFILL_STANDALONE_PCC", "0")),
         ("PREFILL_STANDALONE_CHUNKED_PCC", os.environ.get("PREFILL_STANDALONE_CHUNKED_PCC", "0.88")),
         (
             "PREFILL_STANDALONE_CHUNKED_RECORD_ONLY",
@@ -914,10 +734,7 @@ def main() -> None:
     kv_caches = ADAPTER.allocate_kv_cache(mesh_device=mesh_device, hf_config=hf_config, params=params)
     runtime.compile(kv_caches)
 
-    if os.environ.get("PREFILL_STANDALONE", "0") == "1":
-        _serve_standalone(runtime, kv_caches, mesh_device, hf_config, rank, num_ranks, is_first_rank)
-    else:
-        _serve_request(runtime, kv_caches, mesh_device, hf_config, rank, num_ranks, is_first_rank)
+    _serve_request(runtime, kv_caches, mesh_device, hf_config, rank, num_ranks, is_first_rank)
 
     # Release captured traces + the sub-device managers that own them BEFORE closing the mesh: the
     # trace buffers live inside the MoE-overlap SubDeviceManagers, so closing with both registered
@@ -932,51 +749,9 @@ def main() -> None:
     logger.info(f"[pp rank {rank}] shutdown complete")
 
 
-def _serve_standalone(
-    runtime, kv_caches, mesh_device, hf_config, rank: int, num_ranks: int, is_first_rank: bool
-) -> None:
-    """Bring-up / benchmark path: golden-trace input on rank 0, D2D-socket transport between ranks,
-    per-rank KV PCC. Self-contained (no external producer); covers num_ranks 1..N."""
-    # Warm-up sync — the ONLY barrier. Every rank finishes compile before any chunk enters the
-    # pipeline, so a downstream rank isn't still warming up while an upstream one races ahead. The
-    # per-chunk loop takes no barrier. Trade-off: a rank that dies during compile hangs the others here.
-    ttnn.distributed_context_barrier()
-
-    # D2D transport: with >1 rank, every rank stands up its pipeline endpoints (revert the custom
-    # sub-device as above). The post-compile barrier guarantees all ranks reach the chained create
-    # rendezvous. A single rank owns the whole model — no transport.
-    d2d_in = d2d_out = None
-    if num_ranks > 1:
-        mesh_device.clear_loaded_sub_device_manager()
-        d2d_in, d2d_out = build_d2d_pipeline_endpoints(mesh_device, rank, num_ranks, CHUNK_SIZE, hf_config.hidden_size)
-        # The chained D2D socket rendezvous finishes at staggered times per rank. Without this barrier
-        # rank 0 enters its produce loop first, fills the socket, and stalls ~6s waiting for the
-        # downstream ranks to enter their consume loops — moving that skew out of the timed chunk loop.
-        ttnn.distributed_context_barrier()
-
-    # Capture the trace (use_trace) HERE — after D2D endpoints are built (their receiver-socket L1 must be
-    # allocated before the trace records, or it corrupts replay on the last rank) and before the chunk loop,
-    # so the one-time capture stays out of the timed loop.
-    if getattr(runtime, "capture_trace", None) and runtime.config.use_trace:
-        runtime.capture_trace(kv_caches)
-
-    logger.info(f"[pp rank {rank}] setup complete, entering standalone loop")
-    run_standalone_loop(runtime, kv_caches, rank, num_ranks, d2d_in=d2d_in, d2d_out=d2d_out)
-
-    if d2d_in is not None or d2d_out is not None:
-        # Free the services while the mesh + command queues are still alive (their dtors free a command
-        # queue and service-core L1; running after close_mesh_device aborts with cq_id-out-of-range).
-        import gc
-
-        d2d_in = d2d_out = None
-        gc.collect()
-
-
 def _serve_request(runtime, kv_caches, mesh_device, hf_config, rank: int, num_ranks: int, is_first_rank: bool) -> None:
     """Production serving: token chunks + PrefillMetadata arrive over the H2D socket from an external
-    producer (prefill_producer.py / the scheduler); unbounded (runs to SIGTERM). Same pipeline
-    mechanics as standalone (num_ranks 1..N over D2D); the only difference is the trigger (H2D input)
-    and that it runs forever.
+    producer (prefill_producer.py / the scheduler); unbounded (runs to SIGTERM), num_ranks 1..N over D2D.
 
     Migration (KV-chunk-table publish) runs for any rank count: every rank all-gathers its stage into
     the merged table and rank 0 builds + publishes it. Per-layer completions feed the scheduler channel
@@ -1008,7 +783,7 @@ def _serve_request(runtime, kv_caches, mesh_device, hf_config, rank: int, num_ra
             f"drive it with prefill_producer.py / the scheduler."
         )
 
-    # D2D pipeline transport for num_ranks>1 (same as standalone).
+    # D2D pipeline transport for num_ranks>1.
     d2d_in = d2d_out = None
     if num_ranks > 1:
         mesh_device.clear_loaded_sub_device_manager()
@@ -1378,7 +1153,7 @@ def _serve_request(runtime, kv_caches, mesh_device, hf_config, rank: int, num_ra
     # Capture the trace (use_trace) after the D2D endpoints AND the per-layer completion wiring
     # (LayerAck channel / layer-completion sink) are set up, but before the request loop: the capture
     # must split at each completion point, and doing it here keeps the one-time cost out of the loop.
-    # No-op if already captured. See _serve_standalone / TtPrefillRuntime.capture_trace().
+    # No-op if already captured. See TtPrefillRuntime.capture_trace().
     if getattr(runtime, "capture_trace", None) and runtime.config.use_trace:
         runtime.capture_trace(kv_caches)
 
@@ -1438,7 +1213,7 @@ def _serve_request(runtime, kv_caches, mesh_device, hf_config, rank: int, num_ra
             if is_first_rank and mig_driver is not None:
                 # Interleaved: per-chunk migrates were already issued during the loop; drain the tail
                 # (consume any remaining acks + wait_complete the deferred copies).
-                mig_driver.drain(expected_chunks=NUM_CHUNKS)
+                mig_driver.drain(expected_chunks=CHUNKS_PER_SLOT)
             elif is_first_rank:
                 assert migration_endpoint is not None, "rank 0 must hold the migration client for the self-test"
                 # Loopback target is THIS endpoint's own id (A->B loopback; no peer, no connect_to).
