+++ models/demos/common/prefill/runners/runner_utils.py
@@ -180,3 +180,75 @@ def load_trace_token_ids(trace_dir, total_len=None) -> list:
         md = json.load(f)
     tids = list(md["token_ids"])
     return tids[:total_len] if total_len is not None else tids
+
+
+# ---------------------------------------------------------------------------
+# Layer assignment
+# ---------------------------------------------------------------------------
+
+
+def _snap_counts_to_starts(counts, valid_starts, num_layers):
+    """Nudge an even split's interior rank boundaries onto the nearest valid start (preserving
+    sum == num_layers), for models that constrain where a rank may begin (layer_split_boundaries).
+    Nearest by |distance| then lower index; each boundary is used at most once and stays increasing."""
+    valid = sorted(valid_starts)
+    boundaries, s = [], 0
+    for c in counts[:-1]:
+        s += c
+        boundaries.append(s)
+    snapped, prev = [], 0
+    for b in boundaries:
+        cand = min(
+            (v for v in valid if prev < v < num_layers and v not in snapped),
+            key=lambda v: (abs(v - b), v),
+            default=None,
+        )
+        if cand is None:
+            raise ValueError(f"cannot place {len(counts)} pipeline ranks on valid layer boundaries {valid}")
+        snapped.append(cand)
+        prev = cand
+    out, prev = [], 0
+    for b in [*snapped, num_layers]:
+        out.append(b - prev)
+        prev = b
+    return out
+
+
+def compute_layer_split(num_layers: int, num_ranks: int, valid_starts=None) -> list[tuple[int, int]]:
+    """Contiguous (first_layer_idx, count) per rank. PREFILL_PP_LAYER_COUNTS, a
+    comma-separated count list summing to num_layers, overrides the default even
+    split (remainder handed to the earlier ranks).
+
+    ``valid_starts`` (from the adapter's ``layer_split_boundaries``): layer indices at which a rank may
+    begin. None => unconstrained. When set, the default even split is auto-snapped onto valid
+    boundaries, and any split (explicit or snapped) whose rank starts fall off them is rejected early."""
+    override = os.environ.get("PREFILL_PP_LAYER_COUNTS")
+    if override:
+        counts = [int(x) for x in override.split(",")]
+        if len(counts) != num_ranks or sum(counts) != num_layers:
+            raise ValueError(
+                f"PREFILL_PP_LAYER_COUNTS={override!r} must list {num_ranks} counts summing to "
+                f"{num_layers} (got {len(counts)} counts summing to {sum(counts)})"
+            )
+    else:
+        base, rem = divmod(num_layers, num_ranks)
+        counts = [base + (1 if r < rem else 0) for r in range(num_ranks)]
+        if valid_starts is not None:
+            counts = _snap_counts_to_starts(counts, valid_starts, num_layers)
+
+    ranges = []
+    start = 0
+    for count in counts:
+        ranges.append((start, count))
+        start += count
+
+    if valid_starts is not None:
+        for first_idx, _ in ranges:
+            if first_idx not in valid_starts:
+                near = sorted(b for b in valid_starts if abs(b - first_idx) <= 4)
+                raise ValueError(
+                    f"pipeline rank starts at layer {first_idx}, not a valid boundary for this model "
+                    f"(nearest valid: {near}). Set PREFILL_PP_LAYER_COUNTS so every cumulative boundary "
+                    f"is a valid start."
+                )
+    return ranges
