--- models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_rank_binding_4rank_d2d.yaml
@@ -1,39 +1,0 @@
-# tt-run rank binding for the 4-rank pipeline prefill over the D2D socket transport.
-#
-# Four 8x4 galaxies, one rank each, chained rank 0 -> 1 -> 2 -> 3 by inter-galaxy MeshSockets over the
-# connected 4-galaxy MGD. The model's layers split evenly across the four ranks; the hidden state hops
-# device-to-device at each boundary (no host/NFS round-trip).
-#
-# Launch on the 4-galaxy superpod (model chosen via PREFILL_MANIFEST). Host order must match the physical
-# cabling chain (mesh 0 -> first --host, ... mesh 3 -> last); reorder the --host slots if a boundary's
-# fabric fails to train:
-#   PREFILL_MANIFEST=models/demos/deepseek_v3_d_p/tt/runners/manifests/kimi26.json \
-#     ./models/demos/common/prefill/runners/run_pipeline_prefill.sh \
-#     models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_rank_binding_4rank_d2d.yaml \
-#     bh-glx-d07u02:1,bh-glx-d07u08:1,bh-glx-d08u02:1,bh-glx-d08u08:1
-
-rank_bindings:
-  - rank: 0
-    mesh_id: 0
-    mesh_host_rank: 0
-    env_overrides: {}
-  - rank: 1
-    mesh_id: 1
-    mesh_host_rank: 0
-    env_overrides: {}
-  - rank: 2
-    mesh_id: 2
-    mesh_host_rank: 0
-    env_overrides: {}
-  - rank: 3
-    mesh_id: 3
-    mesh_host_rank: 0
-    env_overrides: {}
-
-# Connected 4x 8x4 BH-galaxy MGD: inter-galaxy links 0<->1<->2<->3.
-mesh_graph_desc_path: models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_4galaxy_connected_mesh_graph_descriptor.textproto
-
-# Topology + transport only — model and run-profile (PCC, chunk count, model knobs) come from PREFILL_MANIFEST.
-global_env:
-  PREFILL_STANDALONE: "1"   # standalone bring-up/pipeline mode, not the production request loop
-  PREFILL_FABRIC_MODE: "2d"  # D2D pipeline transport needs FABRIC_2D
