+++ models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_rank_binding_8rank_d2d.yaml
@@ -52,5 +52,4 @@ mesh_graph_desc_path: models/demos/common/prefill/runners/topology_configuration
 
 # Topology + transport only — model and run-profile (PCC, chunk count, model knobs) come from PREFILL_MANIFEST.
 global_env:
-  PREFILL_STANDALONE: "1"   # standalone bring-up/pipeline mode, not the production request loop
   PREFILL_FABRIC_MODE: "2d"  # D2D pipeline transport needs FABRIC_2D
