--- models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_real_1galaxy_1rank.yaml
@@ -1,30 +1,0 @@
-# tt-run rank binding for a 1-rank REAL-model de-risk on ONE BH galaxy.
-#
-# Single rank, full galaxy: is_first == is_last == True, so the runner builds the WHOLE
-# model (embedding + all layers + norm + LM head) and runs it with NO cross-rank
-# transport. Confirms, in isolation from the 2-host path, that:
-#   * per-galaxy FABRIC_1D trains with the [LINE, RING] 8x4 torus dim_types,
-#   * the model builds from the shared TTNN weight cache, and
-#   * the populated KV cache PCC-matches the golden trace.
-# It is the precursor to the multi-galaxy D2D pipeline split (pipeline_prefill_rank_binding_2rank_d2d.yaml).
-#
-# Uses the in-tree known-good single-galaxy descriptor (8x4, [LINE, RING]); our 2-galaxy
-# descriptor's two meshes are copies of it.
-#
-# Launch (model chosen via PREFILL_MANIFEST):
-#   PREFILL_MANIFEST=models/demos/deepseek_v3_d_p/tt/runners/manifests/deepseek_v3.json \
-#     ./models/demos/common/prefill/runners/run_pipeline_prefill.sh \
-#     models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_real_1galaxy_1rank.yaml \
-#     bh-glx-d03u02:1
-
-rank_bindings:
-  - rank: 0
-    mesh_id: 0
-    mesh_host_rank: 0
-    env_overrides: {}
-
-mesh_graph_desc_path: tt_metal/fabric/mesh_graph_descriptors/single_bh_galaxy_torus_x_graph_descriptor.textproto
-
-# Topology + transport only — model and run-profile (PCC, chunk count, model knobs) come from PREFILL_MANIFEST.
-global_env:
-  PREFILL_STANDALONE: "1"   # standalone (bring-up/pipeline) mode, not the production request loop
