+++ models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_request_2rank.yaml
@@ -1,6 +1,9 @@
-# REQUEST mode (unbounded, H2D-driven), 2-rank D2D pipeline. No PREFILL_STANDALONE => request loop.
-# rank0 (d07u02) builds the H2D service; downstream rank1 reads via D2D. Migration/LayerAck auto-disabled
-# for num_ranks>1. Cache sized to 11 chunks. Shutdown is via SIGKILL (no end-of-request sentinel yet).
+# REQUEST mode (unbounded, H2D-driven), 2-rank D2D pipeline.
+# rank0 (d07u02) builds the H2D service; downstream rank1 reads via D2D. Migration/LayerAck work at this
+# rank count too (every rank all-gathers the KV table, rank0 publishes it). Cache sized to 11 chunks.
+# Shutdown: the producer's all -1 sentinel (PREFILL_SEND_SHUTDOWN=1); SIGKILL is the fallback.
+# Non-default model: add PREFILL_MANIFEST (absolute path) to global_env below — shell-forwarded env
+# reaches only the launch-host rank.
 rank_bindings:
   - {rank: 0, mesh_id: 0, mesh_host_rank: 0, env_overrides: {}}
   - {rank: 1, mesh_id: 1, mesh_host_rank: 0, env_overrides: {}}
