+++ models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_request_4rank.yaml
@@ -1,6 +1,9 @@
-# REQUEST mode (unbounded, H2D-driven), 4-rank D2D pipeline. No PREFILL_STANDALONE => request loop.
-# rank0 builds the H2D service; downstream ranks read via D2D. Migration/LayerAck auto-disabled (num_ranks>1).
-# Shutdown is SIGKILL (no end-of-request sentinel yet). Host order is the ring chain (see the 4-galaxy MGD).
+# REQUEST mode (unbounded, H2D-driven), 4-rank D2D pipeline.
+# rank0 builds the H2D service; downstream ranks read via D2D. Migration/LayerAck work at this rank count
+# too (every rank all-gathers the KV table, rank0 publishes it). Shutdown: the producer's all -1 sentinel
+# (PREFILL_SEND_SHUTDOWN=1); SIGKILL is the fallback. Host order is the ring chain (see the 4-galaxy MGD).
+# Non-default model: add PREFILL_MANIFEST (absolute path) to global_env below — shell-forwarded env
+# reaches only the launch-host rank.
 rank_bindings:
   - {rank: 0, mesh_id: 0, mesh_host_rank: 0, env_overrides: {}}
   - {rank: 1, mesh_id: 1, mesh_host_rank: 0, env_overrides: {}}
