+++ models/demos/deepseek_v3_d_p/tt/runners/manifests/deepseek_v3.json
@@ -1,8 +1,6 @@
 {
   "env": {
     "PREFILL_MODEL": "deepseek_v3_d_p",
-    "PREFILL_GATE_FALLBACK_MODE": "DEVICE_FP32",
-    "PREFILL_STANDALONE_PCC": "1",
-    "PREFILL_STANDALONE_NCHUNKS": "11"
+    "PREFILL_GATE_FALLBACK_MODE": "DEVICE_FP32"
   }
 }
