+++ models/demos/deepseek_v3_d_p/tt/runners/manifests/kimi26.json
@@ -2,8 +2,6 @@
   "env": {
     "PREFILL_MODEL": "kimi_k2_6",
     "PREFILL_PP_D2D_FIFO_BYTES": "4096",
-    "PREFILL_GATE_FALLBACK_MODE": "DEVICE_FP32",
-    "PREFILL_STANDALONE_PCC": "0",
-    "PREFILL_STANDALONE_NCHUNKS": "11"
+    "PREFILL_GATE_FALLBACK_MODE": "DEVICE_FP32"
   }
 }
