+++ models/demos/minimax_m3/docs/PIPELINE_PREFILL_TESTING.md
@@ -1,7 +1,7 @@
 # MiniMax-M3 pipeline prefill — running & testing
 
-Multi-galaxy pipeline-parallel prefill for MiniMax-M3 via the common prefill runner. Accuracy (KV PCC)
-uses the standalone bindings; throughput + overlap plots use request mode + the producer.
+Multi-galaxy pipeline-parallel prefill for MiniMax-M3 via the common prefill runner. Serving is always
+request mode (runner + producer); accuracy (KV PCC) and throughput differ only in the producer flags.
 
 ## Setup
 
@@ -25,31 +25,24 @@ Run the commands below from inside the allocation.
 
 ## Accuracy — KV PCC
 
-Run from `TT_METAL_HOME`. PASS = every rank's KV PCC passes.
+Same two processes as the perf run below (runner under `tt-run`, producer on rank 0's host); the producer
+reads the KV back and PCCs it against the golden trace. Start the runner with the request binding for the
+rank count you want:
 
-### 1 galaxy
 ```bash
 PREFILL_MANIFEST=models/demos/minimax_m3/tt/runners/manifests/minimax_m3.json \
   ./models/demos/common/prefill/runners/run_pipeline_prefill.sh \
-  models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_real_1galaxy_1rank.yaml \
+  models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_request_1rank.yaml \
   bh-glx-b08u02:1
 ```
 
-### 2 galaxies
-```bash
-PREFILL_MANIFEST=models/demos/minimax_m3/tt/runners/manifests/minimax_m3.json \
-  ./models/demos/common/prefill/runners/run_pipeline_prefill.sh \
-  models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_rank_binding_2rank_d2d.yaml \
-  bh-glx-b08u02:1,bh-glx-b08u08:1
-```
+then add `PREFILL_PRODUCER_CHECK_PCC=1` to the producer command (Process 2 below) and drop
+`PREFILL_PRODUCER_MAX_REQUESTS` down to one request per slot so every slot's KV is still resident when it
+is read back. PASS = `[producer] KV cache PCC PASSED` (threshold `PREFILL_STANDALONE_CHUNKED_PCC`,
+default `0.93`).
 
-### 4 galaxies
-```bash
-PREFILL_MANIFEST=models/demos/minimax_m3/tt/runners/manifests/minimax_m3.json \
-  ./models/demos/common/prefill/runners/run_pipeline_prefill.sh \
-  models/demos/common/prefill/runners/topology_configuration/pipeline_prefill_rank_binding_4rank_d2d.yaml \
-  bh-glx-b09u02:1,bh-glx-b09u08:1,bh-glx-b08u08:1,bh-glx-b08u02:1
-```
+Swap `..._request_1rank.yaml` for `..._request_2rank.yaml` / `..._request_4rank.yaml` and extend the
+`--host` list (see the table below) for 2 and 4 galaxies.
 
 ## Perf — throughput + overlap plot
